MergeTreeFileCommitProtocol trait 1
